+++ includes/acl/compression/stream/convert_rotation_streams.h
@@ -68,7 +68,7 @@ namespace acl
 			ACL_ASSERT(bone_stream.rotations.get_sample_size() == sizeof(Quat_32), "Unexpected rotation sample size. %u != %u", bone_stream.rotations.get_sample_size(), sizeof(Quat_32));
 
 			const uint32_t num_samples = bone_stream.rotations.get_num_samples();
-			const uint32_t sample_rate = bone_stream.rotations.get_sample_rate();
+			const float sample_rate = bone_stream.rotations.get_sample_rate();
 			RotationTrackStream converted_stream(allocator, num_samples, sizeof(Quat_32), sample_rate, high_precision_format);
 
 			for (uint32_t sample_index = 0; sample_index < num_samples; ++sample_index)
